# slightly faster USB transport