--- a/pyxcp/__init__.py
+++ b/pyxcp/__init__.py
@@ -31,4 +31,4 @@ __copyright__ = """
 """
 
 # if you update this manually, do not forget to update .bumpversion.cfg
-__version__ = "0.18.41"
+__version__ = "0.18.42"