## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] I have read the **CONTRIBUTING** document.
- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.

```
b = array("B", bytes(4096))
%timeit bytes(b)
244 ns ± 8.3 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
%timeit b.tobytes()
162 ns ± 1.43 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
%timeit bytes(b[:2000])
375 ns ± 14.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
%timeit b.tobytes()[:2000]
293 ns ± 9.4 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```